+++ docs/releases.md
@@ -13,7 +13,8 @@
 - Fixed bug in side bar bottom action not being visible in mobile mode
 - Much improved UI for Seven Wonders games, now also allowing for Military and Scientific victories in the Duel edition
 - Updated documentation
-- Added theme switch action for dark/light mode
+- Added theme switch action for dark/light mode; use system-theme by default
+- Improved sidebar layout
 
 ### v1.0.0 - Initial Release (2025-01-03)
 - First version allowing to create scoreboards for simple games with multiple players.